refactor: BridgeFinder moved to utils
