+++ src/test/kotlin/model/BridgeFinderTest.kt
@@ -1,5 +1,6 @@
 package model
 
+import model.utils.BridgeFinder
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.BeforeEach
